Screen for Interactions and Initialize Matchmaker Object

Screen for interactiins and initialize a Matchmaker object.

Screening(
  data,
  annotation,
  interaction,
  partner_a = 1,
  partner_b = 2,
  path = FALSE,
  project_name = "",
  zero_percent = 0.7
)

Arguments

data

Expression data. A d x M matrix with d rows of features and M columns of data points (cells).

annotation

Cell type identity in following formats:

character, cell type identity with either names as cell ids (barcodes) or in the same order matching the column names in data.

data.frame, cell ids (barcodes) in row names and cell type identity in first column.

interaction

Interaction partner pair data with at least two columns indicating gene-gene interaction partners.

partner_a

Column name or index for interaction partner a. Defaulte is 1, the first column in interaction pair data.

partner_b

Column name or index for interaction partner b. Defaulte is 2, the second column in interaction pairs data.

path

Whether inputs are in directory path format.

project_name

Name of the project.

zero_percent

Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70%.

Value

Returns a Matchmaker object.

Examples

if (FALSE) {
object <- Screening(data = data.use, annotation = idents.use, interaction = interaction.data)
}